requirements.txt
# Core
streamlit==1.53.1
pandas==2.3.1
numpy==2.3.1

# Database
mysql-connector-python==8.2.0
pymysql==1.1.0
SQLAlchemy==2.0.48

# ML
scikit-learn==1.7.1
xgboost==3.2.0
lightgbm==4.6.0
shap==0.51.0

# Visualization
plotly==6.5.2
matplotlib==3.10.5

# Reports
reportlab==4.0.7

# Auth
bcrypt==4.1.2

# Scheduling/GA
DEAP==1.4.1

# Testing
pytest==7.4.3

# Other
python-dotenv==1.0.0
psutil==5.9.6
